requirements.txt
openai
numpy
python-dotenv
fastapi
uvicorn
pydantic
PyYAML
tinydb